+++ lua/community.lua
@@ -128,6 +128,17 @@ return {
   -- Utility
   -- rich command prompt
   { import = "astrocommunity.utility.noice-nvim" },
+  {
+    "noice.nvim",
+    opts = function(_, opts)
+      local utils = require "astrocore"
+      return utils.extend_tbl(opts, {
+        presets = {
+          lsp_doc_border = true, -- add a border to hover docs and signature help
+        },
+      })
+    end,
+  },
   -- Override notification display defaults
   {
     "rcarriga/nvim-notify",