+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - community: local fork to test clojure pack changes
 - community: import clojure pack, add practicalli config overrides
 - github: Octo plugin with <leader>gh keymaps
+- community: `lsp_doc_border = true` adds border to hover docs & signature help
 
 ## Changed
 - readme: expand on description and use